requirements.txt
APScheduler==3.11.0
discord.py==2.5.2
python-dotenv==1.1.0
fastapi
google-api-python-client
google-auth-oauthlib
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
uvicorn
aiohttp
requests
